+++ NEWS.md
@@ -7,6 +7,8 @@
 
 * Fixed issue in `find_formula()` for models of class `glmmPQL` (package *MASS*).
 
+* Fixed issue in `get_variance()` for *glmmTMB* models with `family = "ordbeta"`.
+
 # insight 0.19.7
 
 ## General